separate registration view
