+++ frontend/component/shared/RouteSwitch.tsx
@@ -1,15 +1,20 @@
+import _ from 'rambda'
 import cx from 'classnames'
 import { Link, withRouter } from 'react-router-dom'
+import { connect } from 'redux-zero/react'
 import { Grid } from '@utility'
 import { SITE_MAP } from '@constant'
 
 const cls = active => cx('col-auto route-switch--route m-1', { active, 'drop-shadow': active })
 
-export const RouteSwitch = withRouter(({ history }) => {
+const WrappedRouteSwitch = ({ history, registeredRelayers, currentUser }) => {
   const activeRoute = history.location.pathname
+  const addresses = _.pluck('address', registeredRelayers)
+  const isRegistered = currentUser !== '' && registeredRelayers.length > 0 && currentUser in addresses
+  const availableRoutes = Object.keys(SITE_MAP).filter(r => isRegistered ? r !== 'Registration' : r !== 'Dashboard')
   return (
     <Grid className="route-switch mt-1 mb-1">
-      {Object.keys(SITE_MAP).filter(r => r !== 'Auth').map(route => (
+      {availableRoutes.map(route => (
         <Link
           to={SITE_MAP[route]}
           className={cls(activeRoute === SITE_MAP[route])}
@@ -22,4 +27,16 @@ export const RouteSwitch = withRouter(({ history }) => {
       ))}
     </Grid>
   )
+}
+
+const mapProps = store => ({
+  registeredRelayers: store.relayers,
+  currentUser: store.currentUserAddress,
 })
+
+const connector = _.compose(
+  withRouter,
+  connect(mapProps),
+)
+
+export const RouteSwitch = connector(WrappedRouteSwitch)
