+++ package.json
@@ -40,6 +40,7 @@
     "dotenv": "^6.2.0",
     "ethers": "^4.0.20",
     "formik": "^1.4.2",
+    "rambda": "^2.3.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-hot-loader": "^4.6.3",
